refactor: move code loading to its own module
